+++ src/main/java/thaumicboots/api/ITBootJumpable.java
@@ -1,0 +1,7 @@
+package thaumicboots.api;
+
+public interface ITBootJumpable {
+    double getJumpModifier();
+    void toggle();
+    boolean getToggle();
+}
