+++ src/main/java/thaumicboots/item/boots/unique/ItemMeteoricCometBoots.java
