+++ src/main/java/thaumicboots/item/boots/voidwalker/ItemMeteorVoidwalkerBoots.java
@@ -28,145 +28,23 @@ import thaumcraft.client.fx.ParticleEngine;
 import thaumcraft.client.fx.particles.FXWispEG;
 import thaumcraft.common.Thaumcraft;
 import thaumcraft.common.items.armor.Hover;
+import thaumicboots.api.ItemVoidBoots;
 import thaumicboots.main.utils.TabThaumicBoots;
 
-public class ItemMeteorVoidwalkerBoots extends ItemArmor
+public class ItemMeteorVoidwalkerBoots extends ItemVoidBoots
         implements IVisDiscountGear, IWarpingGear, IRunicArmor, IRepairable, ISpecialArmor {
 
     public ItemMeteorVoidwalkerBoots(final ArmorMaterial material, final int j, final int k) {
         super(material, j, k);
-        setCreativeTab(TabThaumicBoots.tabThaumicBoots);
-        setUnlocalizedName("ItemVoidMeteor");
-        setTextureName("thaumicboots:purpleHaze_16x");
-
-        MinecraftForge.EVENT_BUS.register(this);
     }
 
-    @Override
-    public String getArmorTexture(final ItemStack stack, final Entity entity, final int slot, final String type) {
-        return "thaumicboots:model/VoidwalkerBootsMeteor_-_Purple.png";
-    }
-
-    @Override
-    public EnumRarity getRarity(final ItemStack stack) {
-        return EnumRarity.epic;
-    }
-
-    @Override
-    public int getRunicCharge(final ItemStack stack) {
-        return 0;
-    }
-
-    @Override
-    public int getWarp(final ItemStack stack, final EntityPlayer player) {
-        return 5;
-    }
-
-    @Override
-    public int getVisDiscount(final ItemStack stack, final EntityPlayer player, final Aspect aspect) {
-        return 5;
-    }
-
-    @Override
-    public ArmorProperties getProperties(final EntityLivingBase entity, final ItemStack stack,
-            final DamageSource source, final double dmg, final int slot) {
-        int priority = 0;
-        double ratio = damageReduceAmount / 90.0D;
-
-        if (source.isMagicDamage() || source.isFireDamage() || source.isExplosion()) {
-            priority = 1;
-            ratio = damageReduceAmount / 80.0D;
-        } else if (source.isUnblockable()) {
-            priority = 0;
-            ratio = 0.0D;
-        }
-        return new ArmorProperties(priority, ratio, stack.getMaxDamage() + 1 - stack.getItemDamage());
-    }
-
-    @Override
-    public int getArmorDisplay(final EntityPlayer player, final ItemStack stack, final int slot) {
-        return damageReduceAmount;
-    }
-
-    @Override
-    public void damageArmor(final EntityLivingBase entity, final ItemStack stack, final DamageSource source,
-            final int dmg, final int slot) {
-        if (source != DamageSource.fall) {
-            stack.damageItem(dmg, entity);
-        }
-    }
-
-    @Override
-    public void addInformation(final ItemStack stack, final EntityPlayer player, final List list, final boolean b) {
-        list.add(
-                EnumChatFormatting.DARK_PURPLE + StatCollector.translateToLocal("tc.visdiscount")
-                        + ": "
-                        + getVisDiscount(stack, player, null)
-                        + "%");
-    }
-
-    @Override
-    public void onUpdate(final ItemStack stack, final World world, final Entity entity, final int j, final boolean k) {
-        super.onUpdate(stack, world, entity, j, k);
-        if (!world.isRemote && stack.isItemDamaged()
-                && entity.ticksExisted % 20 == 0
-                && entity instanceof EntityLivingBase) {
-            stack.damageItem(-1, (EntityLivingBase) entity);
-        }
-    }
-
-    @Override
-    public void onArmorTick(final World world, final EntityPlayer player, final ItemStack stack) {
-        super.onArmorTick(world, player, stack);
-        // repair
-        if (!world.isRemote && stack.isItemDamaged() && player.ticksExisted % 20 == 0) {
-            stack.damageItem(-1, player);
-        }
-
-        // particles
-        final double motion = Math.abs(player.motionX) + Math.abs(player.motionZ) + Math.abs(0.5 * player.motionY);
-        if (world.isRemote && (motion > 0.1D || !player.onGround) && world.rand.nextInt(3) == 0) {
-            particles(world, player);
-        }
-
-        if (player.moveForward > 0.0F) {
-            // increased step height
-            if (player.worldObj.isRemote && !player.isSneaking()) {
-                if (!Thaumcraft.instance.entityEventHandler.prevStep.containsKey(player.getEntityId())) {
-                    Thaumcraft.instance.entityEventHandler.prevStep.put(player.getEntityId(), player.stepHeight);
-                }
-                player.stepHeight = 1.0F;
-            }
-
-            // speed boost
-            if (player.onGround || player.capabilities.isFlying) {
-                float bonus = 0.300F;
-                final ItemStack sash = PlayerHandler.getPlayerBaubles(player).getStackInSlot(3);
-                if (sash != null && sash.getItem() == ItemRegistry.ItemVoidwalkerSash) {
-                    bonus *= 3.0F;
-                }
-
-                player.moveFlying(0.0F, 1.0F, player.capabilities.isFlying ? bonus * 0.75F : bonus);
-            } else if (Hover.getHover(player.getEntityId())) {
-                player.jumpMovementFactor = 0.03F;
-            } else {
-                player.jumpMovementFactor = player.isSprinting() ? 0.045F : 0.04F;
-            }
-        }
-        // negate fall damage
-        if (player.fallDistance > 3.0F) {
-            player.fallDistance = 1.0F;
-        }
-    }
-
-    @SideOnly(Side.CLIENT)
-    private void particles(final World world, final EntityPlayer player) {
-        final FXWispEG fx = new FXWispEG(
-                world,
-                player.posX + (Math.random() - Math.random()) * 0.5D,
-                player.boundingBox.minY + 0.05D + (Math.random() - Math.random()) * 0.1D,
-                player.posZ + (Math.random() - Math.random()) * 0.5D,
-                player);
-        ParticleEngine.instance.addEffect(world, fx);
+    protected void setBootsData() {
+        visDiscount = 5;
+        damageAbsorptionRatio = 1.5D;
+        jumpBonus = 0.275D * 3.2;
+        tier = 3;
+        iconResPath = "thaumicboots:purpleHaze_16x";
+        armorResPath = "thaumicboots:model/VoidwalkerBootsMeteor_-_Purple.png";
+        unlocalisedName = "ItemVoidMeteor";
     }
 }
